requirements.txt
fastapi
uvicorn
celery
redis
pdfminer.six
prometheus-client
prometheus-fastapi-instrumentator